+++ R/get_data_object_glm.R
@@ -24,6 +24,9 @@ get_data_object_spglm <- function(formula, family, data, spcov_initial, xcoord, 
     data_sf <- suppressWarnings(sf::st_centroid(data))
     # store as data frame
     data <- sf_to_df(data_sf)
+    if (!missing(xcoord) || !missing(ycoord)) {
+      warning("data is an sf object. Ignoring xcoord and ycoord arguments.", call. = FALSE)
+    }
     ## name xcoord ".xcoord" to be used later
     xcoord <- ".xcoord"
     ## name ycoord ".ycoord" to be used later