+++ NEWS.md
@@ -8,6 +8,7 @@
 
 * Changed the default `size` argument to the local argument in `predict(..., block = TRUE)` and `augment(..., block = TRUE)` from 1000 to 4000. This enhances the block prediction (i.e., Kriging) approximation's accuracy but can slightly increase computational complexity.
 * Enhanced efficiency of block prediction (i.e., Kriging) for large `newdata` objects.
+* Added a warning message that clarifies `xcoord` and `ycoord` are ignored when `data` is an `sf` object.
 * Minor documentation updates.
 
 # spmodel 0.11.1